@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <Import Project="..\..\..\..\core-api.props" />
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
     <DocumentationFile>bin\$(Configuration)\$(TargetFramework)\Identity.API.xml</DocumentationFile>
     <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
     <UserSecretsId>f7d51143-5f51-4aab-97ae-a4bff6b5ca7f</UserSecretsId>
@@ -12,7 +12,7 @@
   
  
   <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="3.1.*" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="5.0.*" />
   </ItemGroup>
 
   <ItemGroup>
