@@ -1,32 +1,39 @@
 ﻿using Bogus;
+
 using FluentAssertions;
+
 using FluentValidation;
-using FluentValidation.Results;
+
 using Identity.DataStores;
 using Identity.DTO;
 using Identity.Objects;
+
 using MedEasy.DAL.EFStore;
 using MedEasy.DAL.Interfaces;
 using MedEasy.IntegrationTests.Core;
+
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+
 using Moq;
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
-using System.Text;
 using System.Threading.Tasks;
+
 using Xunit;
 using Xunit.Abstractions;
 using Xunit.Categories;
+
 using ValidationResult = FluentValidation.Results.ValidationResult;
 
 namespace Identity.Validators.UnitTests
 {
     [UnitTest]
     [Feature("Accounts")]
-    public class NewAccountInfoValidatorTests : IDisposable, IClassFixture<SqliteDatabaseFixture>
+    public class NewAccountInfoValidatorTests : IAsyncLifetime, IClassFixture<SqliteDatabaseFixture>
     {
         private ITestOutputHelper _outputHelper;
         private IUnitOfWorkFactory _uowFactory;
@@ -36,7 +43,7 @@ namespace Identity.Validators.UnitTests
         public NewAccountInfoValidatorTests(ITestOutputHelper outputHelper, SqliteDatabaseFixture databaseFixture)
         {
             _outputHelper = outputHelper;
-            DbContextOptionsBuilder<IdentityContext> dbContextOptionsBuilder = new DbContextOptionsBuilder<IdentityContext>();
+            DbContextOptionsBuilder<IdentityContext> dbContextOptionsBuilder = new();
             dbContextOptionsBuilder.UseSqlite(databaseFixture.Connection);
             _uowFactory = new EFUnitOfWorkFactory<IdentityContext>(dbContextOptionsBuilder.Options, (options) =>
             {
@@ -45,25 +52,19 @@ namespace Identity.Validators.UnitTests
                 return context;
             });
             _loggerMock = new Mock<ILogger<NewAccountInfoValidator>>();
-            
+
             _sut = new NewAccountInfoValidator(_uowFactory, _loggerMock.Object);
         }
 
-        public async void Dispose()
+        public Task InitializeAsync() => Task.CompletedTask;
+
+        public async Task DisposeAsync()
         {
-            using (IUnitOfWork uow =_uowFactory.NewUnitOfWork())
-            {
-                uow.Repository<Account>().Clear();
-                await uow.SaveChangesAsync()
-                    .ConfigureAwait(false);
-            }
-            _uowFactory = null;
-            _outputHelper = null;
-            _loggerMock = null;
-            _sut = null;
+            using IUnitOfWork uow = _uowFactory.NewUnitOfWork();
+            uow.Repository<Account>().Clear();
+            await uow.SaveChangesAsync()
+                     .ConfigureAwait(false);
         }
-
-
 
         [Fact]
         public void IsNewAccountValidator() => typeof(NewAccountInfoValidator).Should()
@@ -73,7 +74,7 @@ namespace Identity.Validators.UnitTests
         {
             get
             {
-                
+                Faker faker = new();
                 yield return new object[]
                 {
                     Enumerable.Empty<Account>(),
@@ -90,7 +91,6 @@ namespace Identity.Validators.UnitTests
                 };
 
                 {
-                    Faker faker = new Faker();
                     Account account = new Account(id: Guid.NewGuid(),
                                                   username: faker.Person.UserName,
                                                   passwordHash: faker.Lorem.Word(),
@@ -115,8 +115,8 @@ namespace Identity.Validators.UnitTests
                         "An account with the same username already exists"
                     };
                 }
+
                 {
-                    Faker faker = new Faker();
                     Account account = new Account
                     (
                         username: "joker",
@@ -146,7 +146,6 @@ namespace Identity.Validators.UnitTests
                 }
 
                 {
-                    Faker faker = new Faker();
                     Account account = new Account
                     (
                         username: "joker",
@@ -155,7 +154,7 @@ namespace Identity.Validators.UnitTests
                         email: "joker@card-city.com",
                         id: Guid.NewGuid()
                     );
-                    
+
                     yield return new object[]
                     {
                         new[] {account},
@@ -187,8 +186,8 @@ namespace Identity.Validators.UnitTests
                             ConfirmPassword = "smile",
                             Username = "capedcrusader"
                         },
-                        (Expression<Func<ValidationResult, bool>>)(vr => !vr.Errors.Any()),
-                        $"Informations are ok"
+                        (Expression<Func<ValidationResult, bool>>)(vr => vr.Errors.Count == 0),
+                        "Informations are ok"
                     };
                 }
             }
@@ -214,6 +213,7 @@ namespace Identity.Validators.UnitTests
                 .ConfigureAwait(false);
 
             // Assert
+            _outputHelper.WriteLine($"Validation results : {vr.Jsonify()}");
             vr.Should()
                 .Match(validationResultExpectation, reason);
 
